refactor: move command module
