+++ src/plugin/commands/scan-reminders.ts
