+++ src/plugin/commands/show-date-chooser.ts
